+++ RAnalogs/RAnEn/R/print.AnEn.R
@@ -29,27 +29,61 @@ print.AnEn <- function (x) {
   
   existed.names <- names(x)
   
-  if ('analogs' %in% names(x)) {
-    empty <- F
-    cat("Member 'analogs': [test station][test time][FLT][member][type]\n")
-    cat(dim(x$analogs))
-    cat(" (value, search station, search observation time)")
+  # Match names starting with analogs. These members are assumed to
+  # be analogs with 5 dimensions.
+  # 
+  matched <- grep(pattern = '^analogs.*', x = existed.names)
+  if (length(matched) != 0) {
+    
+    for (name in existed.names[matched]) {
+      cat(paste("Member '", name, "': [station][test time][FLT][member][value, station, observation time]", sep = ''), '\n')
+      cat(dim(x[[name]]))
+      cat("\n")
+      empty <- F
+    }
+    
+    existed.names <- existed.names[-matched]
     cat("\n")
-    existed.names <- existed.names[-which(existed.names == 'analogs')]
   }
   
-  if ('similarity' %in% names(x)) {
-  	empty <- F
-  	cat("Member 'similarity': [test station][test time][FLT][member][type]\n")
-  	cat(dim(x$similarity))
-  	cat(" (value, search station, search forecast time)")
-  	cat("\n")
-  	existed.names <- existed.names[-which(existed.names == 'similarity')]
+  # Match names starting with similarity. These members are assumed to
+  # be similarity matrices with 5 dimensions.
+  # 
+  matched <- grep(pattern = '^similarity.*', x = existed.names)
+  if (length(matched) != 0) {
+    
+    for (name in existed.names[matched]) {
+      empty <- F
+      cat(paste("Member '", name, "': [station][test time][FLT][member][value, station, forecast time]", sep = ''), '\n')
+      cat(dim(x[[name]]))
+      cat("\n")
+      empty <- F
+    }
+    
+    existed.names <- existed.names[-matched]
+    cat("\n")
+  }
+  
+  # Match names starting with bias. These members are assumed to
+  # be bias matrices with 3 dimensions.
+  # 
+  matched <- grep(pattern = '^bias*', x = existed.names)
+  if (length(matched) != 0) {
+    
+    for (name in existed.names[matched]) {
+      cat(paste("Member '", name, "': [test station][test time][FLT]", sep = ''), '\n')
+      cat(dim(x[[name]]))
+      cat("\n")
+      empty <- F
+    }
+    
+    existed.names <- existed.names[-matched]
+    cat("\n")
   }
   
   if ('mapping' %in% names(x)) {
     empty <- F
-    cat("Member 'mapping': [FLT][forecast time] ")
+    cat("Member 'mapping': [FLT][forecast time]\n")
     cat(dim(x$mapping))
     cat("\n")
     existed.names <- existed.names[-which(existed.names == 'mapping')]
@@ -57,7 +91,7 @@ print.AnEn <- function (x) {
   
   if ('searchStations' %in% names(x)) {
     empty <- F
-    cat("Member 'searchStations': [search station][test station] ")
+    cat("Member 'searchStations': [search station][test station]\n")
     cat(dim(x$searchStations))
     cat("\n")
     existed.names <- existed.names[-which(existed.names == 'searchStations')]
@@ -65,60 +99,19 @@ print.AnEn <- function (x) {
   
   if ('std' %in% names(x)) {
   	empty <- F
-  	cat("Member 'std': [parameter][test station][FLT] ")
+  	cat("Member 'std': [parameter][test station][FLT]\n")
   	cat(dim(x$std))
   	cat("\n")
   	existed.names <- existed.names[-which(existed.names == 'std')]
   }
-  
-  if ('analogs.cor.insitu' %in% names(x)) {
-    empty <- F
-    cat("Member 'analogs.cor.insitu': [test station][test time][FLT][member][type]\n")
-    cat(dim(x$analogs.cor.insitu))
-    cat(" (value, search station, search observation time)")
-    cat("\n")
-    existed.names <- existed.names[-which(existed.names == 'analogs.cor.insitu')]
-  }
-  
-  if ('bias.insitu' %in% names(x)) {
-    empty <- F
-    cat("Member 'bias.insitu': [test station][test time][FLT] ")
-    cat(dim(x$bias.insitu))
-    cat("\n")
-    existed.names <- existed.names[-which(existed.names == 'bias.insitu')]
-  }
-  
-  if ('analogs.cor.nnet' %in% names(x)) {
-    empty <- F
-    cat("Member 'analogs.cor.nnet': [test station][test time][FLT][member][type]\n")
-    cat(dim(x$analogs.cor.nnet))
-    cat(" (value, search station, search observation time)")
-    cat("\n")
-    existed.names <- existed.names[-which(existed.names == 'analogs.cor.nnet')]
-  }
-  
-  if ('bias.nnet' %in% names(x)) {
-    empty <- F
-    cat("Member 'bias.nnet': [test station][test time][FLT] ")
-    cat(dim(x$bias.nnet))
-    cat("\n")
-    existed.names <- existed.names[-which(existed.names == 'bias.nnet')]
-  }
-  
-  if ('bias.model.nnet' %in% names(x)) {
-    empty <- F
-    cat("Member 'bias.model.nnet':  ")
-    cat(class(x$bias.model.nnet))
-    cat("\n")
-    existed.names <- existed.names[-which(existed.names == 'bias.model.nnet')]
-  }
+  cat('\n')
   
   if (empty) {
     cat('[empty list]\n')
   }
   
   if (length(existed.names) != 0) {
-    cat("Some extra member exists:",
-        paste(existed.names, collapse = ', '), '\n')
+    cat("Extra members:\n")
+    cat(paste(existed.names, collapse = ', '), '\n')
   }
 }